# docs/specs/openapi.yaml
openapi: 3.0.3
info:
  title: Oikonomia API
  version: 1.0.0
  description: |
    **Oikonomia** (Credit Card Analyzer) — Personal finance management platform for Argentine users.

    ## Architecture
    - **Backend**: FastAPI (Python) with PostgreSQL, Redis, and Celery
    - **Frontend**: Vite + React SPA
    - **Integrations**: Telegram Bot, WhatsApp Bot, Gemini Flash (AI), Yahoo Finance, IOL/PPI brokers

    ## Key Features
    - Expense & income tracking with encrypted fields (AES-256)
    - Credit/debit card management with billing cycle tracking
    - Hierarchical category system with auto-categorization
    - Budget management with 50/30/20 rule support
    - Investment portfolio tracking with auto price refresh
    - AI-powered spending analysis (Gemini Flash)
    - Smart CSV/PDF bank statement import
    - Family group sharing
    - Real-time SSE notifications
    - Scheduled/recurring expense automation
    - Telegram & WhatsApp bot integrations
    - Admin panel with impersonation and audit logging

    ## Security
    - JWT authentication (access + refresh tokens)
    - TOTP-based MFA
    - Field-level AES-256 encryption
    - HMAC for searchable encrypted fields
    - Rate limiting on sensitive endpoints
  contact:
    name: Oikonomia Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8001
    description: Local development
  - url: https://platform.oikonomia.ar
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /import-jobs:
    get:
      tags: [import-jobs]
      summary: List import jobs
      operationId: listImportJobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PROCESSING, READY_PREVIEW, COMPLETED, FAILED]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Import job list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImportJobResponse'
    post:
      tags: [import-jobs]
      summary: Upload file and create import job
      operationId: createImportJob
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Import job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJobResponse'
  /import-jobs/{job_id}:
    parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [import-jobs]
      summary: Get import job details
      operationId: getImportJob
      responses:
        '200':
          description: Import job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJobResponse'
        '410':
          description: Job expired
    delete:
      tags: [import-jobs]
      summary: Delete import job
      operationId: deleteImportJob
      responses:
        '204':
          description: Import job deleted
  /import-jobs/{job_id}/preview:
    put:
      tags: [import-jobs]
      summary: Update preview data
      operationId: updateImportPreview
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Preview updated
  /import-jobs/{job_id}/confirm:
    post:
      tags: [import-jobs]
      summary: Confirm import and create expenses
      operationId: confirmImportJob
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportConfirmRequest'
      responses:
        '200':
          description: Import completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: integer
                  skipped:
                    type: integer

  # ─────────────────────────────────────────────────────────────
  # DASHBOARD
  # ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ImportJobResponse:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        status:
          type: string
          enum: [PROCESSING, READY_PREVIEW, COMPLETED, FAILED]
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        error_message:
          type: string
          nullable: true
        preview_data:
          type: object
          nullable: true
    ImportConfirmRequest:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            type: object
        cards_mapping:
          type: object
          nullable: true
...